+++ src/main/java/chess/SpringChessApplication.java
@@ -2,24 +2,11 @@ package chess;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @SpringBootApplication
-@Controller
 public class SpringChessApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(SpringChessApplication.class, args);
-    }
-
-    @GetMapping("/")
-    public String index() {
-        return "index.html";
-    }
-
-    @GetMapping("/room/**")
-    public String enter() {
-        return "/game.html";
     }
 }
