+++ src/main/java/chess/controller/SpringChessController.java
@@ -1,0 +1,18 @@
+package chess.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class SpringChessController {
+
+    @GetMapping("/")
+    public String index() {
+        return "/index.html";
+    }
+
+    @GetMapping("/room/**")
+    public String enter() {
+        return "/game.html";
+    }
+}
